+++ test/integration/edge-runtime-module-errors/test/module-imports.test.js
@@ -6,6 +6,7 @@ import { join } from 'path'
 import {
   check,
   fetchViaHTTP,
+  File,
   findPort,
   killApp,
   launchApp,
@@ -13,8 +14,6 @@ import {
   nextStart,
 } from 'next-test-utils'
 import {
-  appOption,
-  context,
   expectModuleNotFoundDevError,
   expectModuleNotFoundProdError,
   expectNoError,
@@ -25,6 +24,25 @@ import {
 
 jest.setTimeout(1000 * 60 * 2)
 
+const context = {
+  appDir: join(__dirname, '../'),
+  logs: { output: '', stdout: '', stderr: '' },
+  api: new File(join(__dirname, '../pages/api/route.js')),
+  lib: new File(join(__dirname, '../lib.js')),
+  middleware: new File(join(__dirname, '../middleware.js')),
+  page: new File(join(__dirname, '../pages/index.js')),
+}
+const appOption = {
+  env: { __NEXT_TEST_WITH_DEVTOOL: 1 },
+  onStdout(msg) {
+    context.logs.output += msg
+    context.logs.stdout += msg
+  },
+  onStderr(msg) {
+    context.logs.output += msg
+    context.logs.stderr += msg
+  },
+}
 const routeUrl = '/api/route'
 const middlewareUrl = '/'
 
@@ -44,20 +62,19 @@ describe('Edge runtime code with imports', () => {
     context.lib.restore()
     context.page.restore()
   })
-
   describe.each([
     {
       title: 'Edge API',
       url: routeUrl,
       init(importStatement) {
         context.api.write(`
-          import { NextResponse } from 'next/server'
-  
+          ${importStatement}
+
           export default async function handler(request) {
-            const { writeFile } = ${importStatement}
-            return Response.json({ ok: writeFile() })
+            basename()
+            return Response.json({ ok: basename() })
           }
-  
+
           export const config = { runtime: 'edge' }
         `)
       },
@@ -68,17 +85,18 @@ describe('Edge runtime code with imports', () => {
       init(importStatement) {
         context.middleware.write(`
           import { NextResponse } from 'next/server'
-  
+          ${importStatement}
+
           export async function middleware(request) {
-            const { writeFile } = ${importStatement}
+            basename()
             return NextResponse.next()
           }
         `)
       },
     },
-  ])('$title dynamically importing node.js module', ({ init, url }) => {
-    const moduleName = 'fs'
-    const importStatement = `await import("${moduleName}")`
+  ])('$title statically importing node.js module', ({ init, url }) => {
+    const moduleName = 'path'
+    const importStatement = `import { basename } from "${moduleName}"`
 
     beforeEach(() => init(importStatement))
 
@@ -114,91 +132,8 @@ describe('Edge runtime code with imports', () => {
       url: routeUrl,
       init(importStatement) {
         context.api.write(`
-          import throwAsync from '../../lib'
-  
           export default async function handler(request) {
-            return Response.json({ ok: await throwAsync() })
-          }
-  
-          export const config = { runtime: 'edge' }
-        `)
-      },
-    },
-    {
-      title: 'Middleware',
-      url: middlewareUrl,
-      init(importStatement) {
-        context.middleware.write(`
-          import { NextResponse } from 'next/server'
-          import throwAsync from './lib'
-  
-          export async function middleware(request) {
-            await throwAsync()
-            return NextResponse.next()
-          }
-        `)
-      },
-    },
-  ])(
-    '$title dynamically importing node.js module in a lib',
-    ({ init, url }) => {
-      const moduleName = 'os'
-      const importStatement = `await import("${moduleName}")`
-
-      beforeEach(() => {
-        init(importStatement)
-        context.lib.write(`
-        export default async function throwAsync() {
-          (${importStatement}).cwd()
-        }
-      `)
-      })
-
-      it('throws unsupported module error in dev at runtime and highlights the faulty line', async () => {
-        context.app = await launchApp(
-          context.appDir,
-          context.appPort,
-          appOption
-        )
-        const res = await fetchViaHTTP(context.appPort, url)
-        expect(res.status).toBe(500)
-        await check(async () => {
-          expectUnsupportedModuleDevError(
-            moduleName,
-            importStatement,
-            await res.text()
-          )
-          return 'success'
-        }, 'success')
-      })
-
-      it('throws unsupported module error in production at runtime and prints error on logs', async () => {
-        const { stderr } = await nextBuild(context.appDir, undefined, {
-          stderr: true,
-        })
-        expect(stderr).toContain(getUnsupportedModuleWarning(moduleName))
-        context.app = await nextStart(
-          context.appDir,
-          context.appPort,
-          appOption
-        )
-        const res = await fetchViaHTTP(context.appPort, url)
-        expect(res.status).toBe(500)
-        expectUnsupportedModuleProdError(moduleName)
-      })
-    }
-  )
-
-  describe.each([
-    {
-      title: 'Edge API',
-      url: routeUrl,
-      init(importStatement) {
-        context.api.write(`
-          ${importStatement}
-  
-          export default async function handler(request) {
-            new Unknown()
+            new (${importStatement})()
             return Response.json({ ok: true })
           }
   
@@ -212,18 +147,17 @@ describe('Edge runtime code with imports', () => {
       init(importStatement) {
         context.middleware.write(`
           import { NextResponse } from 'next/server'
-          ${importStatement}
   
           export async function middleware(request) {
-            new Unknown()
+            new (${importStatement})()
             return NextResponse.next()
           }
         `)
       },
     },
-  ])('$title statically importing 3rd party module', ({ init, url }) => {
+  ])('$title dynamically importing 3rd party module', ({ init, url }) => {
     const moduleName = 'not-exist'
-    const importStatement = `import Unknown from "${moduleName}"`
+    const importStatement = `await import("${moduleName}")`
 
     beforeEach(() => init(importStatement))
 
@@ -242,7 +176,7 @@ describe('Edge runtime code with imports', () => {
       }, 'success')
     })
 
-    it('does not build and reports', async () => {
+    it('does not build and reports module not found error', async () => {
       const { code, stderr } = await nextBuild(context.appDir, undefined, {
         ignoreFail: true,
         stdout: true,
@@ -259,11 +193,11 @@ describe('Edge runtime code with imports', () => {
       url: routeUrl,
       init(importStatement) {
         context.api.write(`
-          ${importStatement}
           export default async function handler(request) {
-            const response = Response.json({ ok: true })
-            response.headers.set('x-from-runtime', nanoid())
-            return response
+            if (process.env === 'production') {
+              new (${importStatement})()
+            }
+            return Response.json({ ok: true })
           }
   
           export const config = { runtime: 'edge' }
@@ -276,40 +210,45 @@ describe('Edge runtime code with imports', () => {
       init(importStatement) {
         context.middleware.write(`
           import { NextResponse } from 'next/server'
-          ${importStatement}
-
+  
           export async function middleware(request) {
-            const response = NextResponse.next()
-            response.headers.set('x-from-runtime', nanoid())
-            return response
+            if (process.env === 'production') {
+              new (${importStatement})()
+            }
+            return NextResponse.next()
           }
         `)
       },
     },
-  ])('$title importing vanilla 3rd party module', ({ init, url }) => {
-    const moduleName = 'nanoid'
-    const importStatement = `import { nanoid } from "${moduleName}"`
+  ])('$title importing unused 3rd party module', ({ init, url }) => {
+    const moduleName = 'not-exist'
+    const importStatement = `await import("${moduleName}")`
 
     beforeEach(() => init(importStatement))
 
-    it('does not throw in dev at runtime', async () => {
+    it('throws not-found module error in dev at runtime and highlights the faulty line', async () => {
       context.app = await launchApp(context.appDir, context.appPort, appOption)
       const res = await fetchViaHTTP(context.appPort, url)
-      expect(res.status).toBe(200)
-      expect(res.headers.get('x-from-runtime')).toBeDefined()
-      expectNoError(moduleName)
+      expect(res.status).toBe(500)
+      await check(async () => {
+        expectModuleNotFoundDevError(
+          moduleName,
+          importStatement,
+          await res.text()
+        )
+        return 'success'
+      }, 'success')
     })
 
-    it('does not throw in production at runtime', async () => {
-      const { stderr } = await nextBuild(context.appDir, undefined, {
+    it('does not build and reports module not found error', async () => {
+      const { code, stderr } = await nextBuild(context.appDir, undefined, {
+        ignoreFail: true,
+        stdout: true,
         stderr: true,
       })
-      expect(stderr).not.toContain(getUnsupportedModuleWarning(moduleName))
-      context.app = await nextStart(context.appDir, context.appPort, appOption)
-      const res = await fetchViaHTTP(context.appPort, url)
-      expect(res.status).toBe(200)
-      expect(res.headers.get('x-from-runtime')).toBeDefined()
-      expectNoError(moduleName)
+      expect(code).toEqual(1)
+
+      expectModuleNotFoundProdError(moduleName, stderr)
     })
   })
 
@@ -319,12 +258,11 @@ describe('Edge runtime code with imports', () => {
       url: routeUrl,
       init(importStatement) {
         context.api.write(`
-          ${importStatement}
-
           export default async function handler(request) {
-            const response = Response.json({ ok: true })
-            response.headers.set('x-from-runtime', Buffer.isBuffer('a string'))
-            return response
+            if (process.env === 'production') {
+              (${importStatement}).spawn('ls', ['-lh', '/usr'])
+            }
+            return Response.json({ ok: true })
           }
   
           export const config = { runtime: 'edge' }
@@ -337,19 +275,19 @@ describe('Edge runtime code with imports', () => {
       init(importStatement) {
         context.middleware.write(`
           import { NextResponse } from 'next/server'
-          ${importStatement}
-
+  
           export async function middleware(request) {
-            const response = NextResponse.next()
-            response.headers.set('x-from-runtime', Buffer.isBuffer('a string'))
-            return response
+            if (process.env === 'production') {
+              (${importStatement}).spawn('ls', ['-lh', '/usr'])
+            }
+            return NextResponse.next()
           }
         `)
       },
     },
-  ])('$title using Buffer polyfill', ({ init, url }) => {
-    const moduleName = 'buffer'
-    const importStatement = `import { Buffer } from "${moduleName}"`
+  ])('$title importing unused node.js module', ({ init, url }) => {
+    const moduleName = 'child_process'
+    const importStatement = `await import("${moduleName}")`
 
     beforeEach(() => init(importStatement))
 
@@ -357,16 +295,17 @@ describe('Edge runtime code with imports', () => {
       context.app = await launchApp(context.appDir, context.appPort, appOption)
       const res = await fetchViaHTTP(context.appPort, url)
       expect(res.status).toBe(200)
-      expect(res.headers.get('x-from-runtime')).toBe('false')
       expectNoError(moduleName)
     })
 
     it('does not throw in production at runtime', async () => {
-      await nextBuild(context.appDir, undefined, { stderr: true })
+      const { stderr } = await nextBuild(context.appDir, undefined, {
+        stderr: true,
+      })
+      expect(stderr).toContain(getUnsupportedModuleWarning(moduleName))
       context.app = await nextStart(context.appDir, context.appPort, appOption)
       const res = await fetchViaHTTP(context.appPort, url)
       expect(res.status).toBe(200)
-      expect(res.headers.get('x-from-runtime')).toBe('false')
       expectNoError(moduleName)
     })
   })
