+++ test/integration/scss-modules/test/dynamic-route.test.js
@@ -1,22 +1,29 @@
 /* eslint-env jest */
 
 import { readdir, readFile, remove } from 'fs-extra'
-import { nextBuild } from 'next-test-utils'
+import { findPort, killApp, nextBuild, nextStart } from 'next-test-utils'
+import webdriver from 'next-webdriver'
 import { join } from 'path'
 
-const fixturesDir = join(__dirname, '../fixtures')
+const fixturesDir = join(__dirname, '../../scss-fixtures')
 
-describe('Custom Properties: Pass-Through IE11', () => {
-  const appDir = join(fixturesDir, 'cp-ie-11')
+describe('Dynamic Route CSS Module Usage', () => {
+  const appDir = join(fixturesDir, 'dynamic-route-module')
 
   let stdout
   let code
+  let app
+  let appPort
+
   beforeAll(async () => {
     await remove(join(appDir, '.next'))
     ;({ code, stdout } = await nextBuild(appDir, [], {
       stdout: true,
     }))
+    appPort = await findPort()
+    app = await nextStart(appDir, appPort)
   })
+  afterAll(() => killApp(app))
 
   it('should have compiled successfully', () => {
     expect(code).toBe(0)
@@ -33,26 +40,52 @@ describe('Custom Properties: Pass-Through IE11', () => {
     const cssContent = await readFile(join(cssFolder, cssFiles[0]), 'utf8')
 
     expect(cssContent.replace(/\/\*.*?\*\//g, '').trim()).toMatchInlineSnapshot(
-      `":root{--color:red}h1{color:var(--color)}"`
+      `"._post__home__a9vTy{background:red}"`
     )
+  })
+
+  it('should apply styles correctly', async () => {
+    const browser = await webdriver(appPort, '/post-1')
+
+    const background = await browser
+      .elementByCss('#my-div')
+      .getComputedCss('background-color')
+
+    expect(background).toMatch(/rgb(a|)\(255, 0, 0/)
   })
 })
 
-describe('Custom Properties: Pass-Through Modern', () => {
-  const appDir = join(fixturesDir, 'cp-modern')
+describe('Catch-all Route CSS Module Usage', () => {
+  const appDir = join(fixturesDir, 'catch-all-module')
 
   let stdout
   let code
+  let app
+  let appPort
+
   beforeAll(async () => {
     await remove(join(appDir, '.next'))
     ;({ code, stdout } = await nextBuild(appDir, [], {
       stdout: true,
     }))
+    appPort = await findPort()
+    app = await nextStart(appDir, appPort)
   })
+  afterAll(() => killApp(app))
 
   it('should have compiled successfully', () => {
     expect(code).toBe(0)
     expect(stdout).toMatch(/Compiled successfully/)
+  })
+
+  it('should apply styles correctly', async () => {
+    const browser = await webdriver(appPort, '/post-1')
+
+    const background = await browser
+      .elementByCss('#my-div')
+      .getComputedCss('background-color')
+
+    expect(background).toMatch(/rgb(a|)\(255, 0, 0/)
   })
 
   it(`should've emitted a single CSS file`, async () => {
@@ -65,39 +98,7 @@ describe('Custom Properties: Pass-Through Modern', () => {
     const cssContent = await readFile(join(cssFolder, cssFiles[0]), 'utf8')
 
     expect(cssContent.replace(/\/\*.*?\*\//g, '').trim()).toMatchInlineSnapshot(
-      `":root{--color:red}h1{color:var(--color)}"`
-    )
-  })
-})
-
-describe('Inline Comments: Minify', () => {
-  const appDir = join(fixturesDir, 'inline-comments')
-
-  let stdout
-  let code
-  beforeAll(async () => {
-    await remove(join(appDir, '.next'))
-    ;({ code, stdout } = await nextBuild(appDir, [], {
-      stdout: true,
-    }))
-  })
-
-  it('should have compiled successfully', () => {
-    expect(code).toBe(0)
-    expect(stdout).toMatch(/Compiled successfully/)
-  })
-
-  it(`should've emitted a single CSS file`, async () => {
-    const cssFolder = join(appDir, '.next/static/css')
-
-    const files = await readdir(cssFolder)
-    const cssFiles = files.filter((f) => /\.css$/.test(f))
-
-    expect(cssFiles.length).toBe(1)
-    const cssContent = await readFile(join(cssFolder, cssFiles[0]), 'utf8')
-
-    expect(cssContent.replace(/\/\*.*?\*\//g, '').trim()).toMatchInlineSnapshot(
-      `"*{box-sizing:border-box}"`
+      `".___post__home__w1yuY{background:red}"`
     )
   })
 })
