+++ test/integration/telemetry/test/page-features.test.js
@@ -56,12 +56,18 @@ describe('page features telemetry', () => {
       if (app) {
         await killApp(app)
       }
-      const event1 = /NEXT_CLI_SESSION_STARTED[\s\S]+?{([\s\S]+?)}/
-        .exec(stderr)
-        .pop()
 
-      expect(event1).toMatch(/"pagesDir": true/)
-      expect(event1).toMatch(/"turboFlag": true/)
+      try {
+        const event1 = /NEXT_CLI_SESSION_STARTED[\s\S]+?{([\s\S]+?)}/
+          .exec(stderr)
+          .pop()
+
+        expect(event1).toMatch(/"pagesDir": true/)
+        expect(event1).toMatch(/"turboFlag": true/)
+      } catch (err) {
+        require('console').error(stderr)
+        throw err
+      }
     } finally {
       await teardown()
     }
@@ -191,24 +197,29 @@ describe('page features telemetry', () => {
         env: { NEXT_TELEMETRY_DEBUG: 1 },
       })
 
-      const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
-        .exec(stderr)
-        .pop()
-      expect(event1).toMatch(/"staticPropsPageCount": 2/)
-      expect(event1).toMatch(/"serverPropsPageCount": 2/)
-      expect(event1).toMatch(/"ssrPageCount": 3/)
-      expect(event1).toMatch(/"staticPageCount": 4/)
-      expect(event1).toMatch(/"totalPageCount": 11/)
-      expect(event1).toMatch(/"totalAppPagesCount": 4/)
-      expect(event1).toMatch(/"serverAppPagesCount": 2/)
-      expect(event1).toMatch(/"edgeRuntimeAppCount": 1/)
-      expect(event1).toMatch(/"edgeRuntimePagesCount": 2/)
+      try {
+        const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
+          .exec(stderr)
+          .pop()
+        expect(event1).toMatch(/"staticPropsPageCount": 2/)
+        expect(event1).toMatch(/"serverPropsPageCount": 2/)
+        expect(event1).toMatch(/"ssrPageCount": 3/)
+        expect(event1).toMatch(/"staticPageCount": 4/)
+        expect(event1).toMatch(/"totalPageCount": 11/)
+        expect(event1).toMatch(/"totalAppPagesCount": 4/)
+        expect(event1).toMatch(/"serverAppPagesCount": 2/)
+        expect(event1).toMatch(/"edgeRuntimeAppCount": 1/)
+        expect(event1).toMatch(/"edgeRuntimePagesCount": 2/)
 
-      const event2 = /NEXT_BUILD_COMPLETED[\s\S]+?{([\s\S]+?)}/
-        .exec(stderr)
-        .pop()
+        const event2 = /NEXT_BUILD_COMPLETED[\s\S]+?{([\s\S]+?)}/
+          .exec(stderr)
+          .pop()
 
-      expect(event2).toMatch(/"totalAppPagesCount": 4/)
+        expect(event2).toMatch(/"totalAppPagesCount": 4/)
+      } catch (err) {
+        require('console').error(stderr)
+        throw err
+      }
     } finally {
       await teardown()
     }
@@ -249,10 +260,15 @@ describe('page features telemetry', () => {
       path.join(appDir, 'pages', '_app_withreportwebvitals.empty')
     )
 
-    const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
-      .exec(build.stderr)
-      .pop()
-    expect(event1).toMatch(/hasReportWebVitals.*?true/)
+    try {
+      const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
+        .exec(build.stderr)
+        .pop()
+      expect(event1).toMatch(/hasReportWebVitals.*?true/)
+    } catch (err) {
+      require('console').error(build.stderr)
+      throw err
+    }
   })
 
   it('detect without reportWebVitals correctly for `next build`', async () => {
@@ -277,9 +293,14 @@ describe('page features telemetry', () => {
       path.join(appDir, 'pages', '_app_withoutreportwebvitals.empty')
     )
 
-    const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
-      .exec(build.stderr)
-      .pop()
-    expect(event1).toMatch(/hasReportWebVitals.*?false/)
+    try {
+      const event1 = /NEXT_BUILD_OPTIMIZED[\s\S]+?{([\s\S]+?)}/
+        .exec(build.stderr)
+        .pop()
+      expect(event1).toMatch(/hasReportWebVitals.*?false/)
+    } catch (err) {
+      require('console').error(build.stderr)
+      throw err
+    }
   })
 })
